+++ Game Tech 2014/nclgl/common.h
@@ -1,12 +1,12 @@
 /******************************************************************************
 Author:Rich Davison
-Description: Some random variables and functions, for lack of a better place 
+Description: Some random variables and functions, for lack of a better place
 to put them.
 
--_-_-_-_-_-_-_,------,   
+-_-_-_-_-_-_-_,------,
 _-_-_-_-_-_-_-|   /\_/\   NYANYANYAN
 -_-_-_-_-_-_-~|__( ^ .^) /
-_-_-_-_-_-_-_-""  ""   
+_-_-_-_-_-_-_-""  ""
 
 *//////////////////////////////////////////////////////////////////////////////
 
@@ -16,19 +16,19 @@ _-_-_-_-_-_-_-""  ""
 #define WEEK_3_CODE
 
 //It's pi(ish)...
-static const float		PI = 3.14159265358979323846f;	
+static const float		PI = 3.14159265358979323846f;
 
 //It's pi...divided by 360.0f!
 static const float		PI_OVER_360 = PI / 360.0f;
 
 //Radians to degrees
-static inline double RadToDeg(const double deg)	{
-	return deg * 180.0 / PI;
+static inline float RadToDeg(const float deg)	{
+	return deg * 180.0f / PI;
 };
 
 //Degrees to radians
-static inline double DegToRad(const double rad)	{
-	return rad * PI / 180.0;
+static inline float DegToRad(const float rad)	{
+	return rad * PI / 180.0f;
 };
 
 //I blame Microsoft...
